Add Flask Web App for Music Recommendations
Updated the project to include a Flask web application for music recommendations. Previously, the recommendations were displayed solely as matplotlib images. Now, users can interact with the application through a user-friendly interface to select their favorite artist and song, receiving personalized music recommendations based on their preferences. The application utilizes Flask for the backend server, JavaScript for asynchronous communication with the server, and HTML/CSS for structuring and styling the frontend interface. This update enhances the overall user experience and provides a more dynamic way to explore music recommendations.